Projects/nbkr_compliance_agent_v2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
attrs==25.3.0
av==14.4.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.6.0b2
bcrypt==4.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==45.0.7
cssselect==1.3.0
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.12.0
fastapi==0.111.0
fastapi-cli==0.0.11
feedparser==6.0.11
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
google-crc32c==1.7.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.34.4
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
lxml==6.0.1
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
openai==1.107.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.2.2
parse==1.20.2
passlib==1.7.4
pathable==0.4.4
pdfminer.six==20231228
pillow==10.4.0
prance==25.4.8.0
propcache==0.3.2
pyasn1==0.6.1
pybars4==0.9.13
pycparser==2.23
pydantic==2.8.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.20.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
pypdf==4.3.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
readability-lxml==0.8.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-toolkit==0.15.1
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.16.2
semantic-kernel==1.29.0
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.1.2
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.1
#uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
python-docx==0.8.11

PyMuPDF==1.24.11
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0
